レビュー修正１

+++ app/controllers/items_controller.rb
@@ -21,8 +21,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to action: :index
+    if current_user.id == item.user.id
+      item.destroy
+      redirect_to action: :index
+    end
   end
 
   def show
